+++ src/Rule/Path.php
@@ -6,12 +6,12 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Router\Matcher;
+namespace Aura\Router\Rule;
 
 use Aura\Router\Route;
 use Psr\Http\Message\ServerRequestInterface;
 
-class Path implements MatcherInterface
+class Path implements RuleInterface
 {
     /**
      *
