+++ src/Rule/RuleInterface.php
@@ -1,10 +1,10 @@
 <?php
-namespace Aura\Router\Matcher;
+namespace Aura\Router\Rule;
 
 use Aura\Router\Route;
 use Psr\Http\Message\ServerRequestInterface;
 
-interface MatcherInterface
+interface RuleInterface
 {
     public function __invoke(ServerRequestInterface $request, Route $route);
 }
